[Chore]#74 성공,에러타입 추가
