+++ main/src/main/java/com/coggiri/main/commons/Enums/ErrorType.java
@@ -21,11 +21,17 @@ public enum ErrorType {
     INVALID_USER_LOGIN_REQUEST(HttpStatus.BAD_REQUEST,4007,"존재하지 않는 유저의 로그인 요청입니다."),
     INVALID_USER_CREATE_ID(HttpStatus.BAD_REQUEST,4008,"이미 존재하는 사용자 아이디입니다."),
     INVALID_JWT_REQUEST(HttpStatus.BAD_REQUEST,4009,"토큰이 존재하지 않습니다"),
+    INVALID_EMAIL_USED(HttpStatus.BAD_REQUEST,4010,"이미 인증에 사용된 이메일입니다."),
+    INVALID_EMAIL_NOT_USED(HttpStatus.BAD_REQUEST,4011,"인증번호를 발급받지 않은 이메일입니다."),
+    INVALID_EMAIL_ACCESS_FREQUENT(HttpStatus.BAD_REQUEST,4012,"비정상적으로 잦은 접근 시도입니다."),
+    INVALID_EMAIL_AUTH_CODE_DIFFERENT(HttpStatus.BAD_REQUEST,4013,"발급된 인증번호와 다릅니다."),
+    INVALID_ID_DUPLICATE(HttpStatus.BAD_REQUEST,4014,"이미 존재하는 아이디입니다."),
     /*
     *  401 Unauthorized (4100 ~ 4199)
     */
     UNAUTHORIZED(HttpStatus.UNAUTHORIZED, 4100, "인증되지 않았습니다."),
-    ACCESS_DENIED(HttpStatus.UNAUTHORIZED, 4109, "알 수 없는 이유로 요청이 거절되었습니다."),
+    ACCESS_DENIED(HttpStatus.UNAUTHORIZED, 4101, "알 수 없는 이유로 요청이 거절되었습니다."),
+    UNAUTHORIZED_EMAIL_EXPIRED(HttpStatus.UNAUTHORIZED,4102,"만료된 인증번호입니다."),
 
     /*
      * 403 Forbidden (4300 ~ 4399)
@@ -45,7 +51,10 @@ public enum ErrorType {
     /*
      * 500 Internal_SERVER (5000 ~ 5099)
      */
-    INTERNAL_SERVER(HttpStatus.INTERNAL_SERVER_ERROR, 5000, "알 수 없는 서버 에러가 발생했습니다");
+    INTERNAL_SERVER(HttpStatus.INTERNAL_SERVER_ERROR, 5000, "알 수 없는 서버 에러가 발생했습니다"),
+    INTERNAL_MAIL_SEND(HttpStatus.INTERNAL_SERVER_ERROR, 5001,"인증 메일을 전송하는데 에러가 발생했습니다.");
+
+
     private final HttpStatus httpStatus;
     private final int code;
     private final String message;
