+++ main/src/main/java/com/coggiri/main/commons/Enums/SuccessType.java
@@ -14,6 +14,10 @@ public enum SuccessType {
     OK(HttpStatus.OK,2000,"성공"),
     SUCCESS_USER_LOGIN(HttpStatus.OK,2001,"로그인에 성공했습니다."),
     SUCCESS_USER_DELETE(HttpStatus.OK,2002,"회원탈퇴에 성공했습니다."),
+    SUCCESS_USER_PASSWORD_CHANGE(HttpStatus.OK,2003,"비밀번호 변경에 성공했습니다"),
+    SUCCESS_EMAIL_SEND(HttpStatus.OK,2004,"인증번호 발급에 성공했습니다."),
+    SUCCESS_EMAIL_VERIFY(HttpStatus.OK,2005,"인증 성공했습니다."),
+    SUCCESS_VALIDATE_ID(HttpStatus.OK,2006,"아이디 중복검사 성공했습니다"),
     /*
     *  201 create (2100 ~ 2199)
     * */
